fix: visual editor preview
